Baro: Wire in drift correction from GPS samples in AHRS.
Give the barometer's drift correction code the GPS altitude at each
gps update.  This allows the baro to estimate how far it has drifted
over the long term.

This assumes that the home altitude is the gps altitude of the point
at which the baro was calibrated.  This is true if the home location
is set at arming/startup, if people give a new home with a correct
relative height (e.g. 0) or the absolute height of the calibrate point
(not the new home location).

The AHRS is starting to feel a bit kitchen-sink-ish but currently
it's the only pose provider that's enabled for all frame types on
all boards.